Add requestMarket action, add getMarketById selector PM-616
