+++ src/routes/MarketDetails/api/index.js
@@ -3,5 +3,4 @@ export { default as sellShares } from './sellShares'
 export { default as fetchMarketTrades } from './fetchMarketTrades'
 export { default as fetchMarketTradesForAccount } from './fetchMarketTradesForAccount'
 export { default as fetchMarketSharesForAccount } from './fetchMarketSharesForAccount'
-export { default as fetchMarket } from './fetchMarket'
 export * from './calculateGasCost'
