+++ src/routes/MarketDetails/store/actions/requestMarket.js
@@ -1,13 +1,29 @@
-import { receiveEntities } from 'store/actions/entities'
-import { fetchMarket } from '../../api'
+import { requestFromRestAPI } from 'api/utils/fetch'
+import addMarkets from 'store/actions/market/addMarkets'
+import { getCollateralToken } from 'store/selectors/blockchain'
+import { hexWithoutPrefix } from 'utils/helpers'
+import { extractMarkets } from 'store/actions/market'
 
-/**
- * Requests details about a single market from GnosisDB.
- * @param {string} marketAddress - Markets Address
- */
-const requestMarket = marketAddress => async (dispatch) => {
-  const payload = await fetchMarket(marketAddress)
-  return dispatch(receiveEntities(payload))
+export const processMarketResponse = (dispatch, state, response) => {
+  if (!response) {
+    dispatch(addMarkets([]))
+    return
+  }
+
+  const applicationCollateralToken = getCollateralToken(state)
+
+  let marketRecords = extractMarkets([response])
+
+  if (applicationCollateralToken.address) {
+    marketRecords = marketRecords.filter(({ collateralToken }) => collateralToken === hexWithoutPrefix(applicationCollateralToken.address))
+  }
+
+  dispatch(addMarkets(marketRecords))
 }
 
-export default requestMarket
+export default marketAddress => async (dispatch, getState) => {
+  const response = await requestFromRestAPI(`markets/${marketAddress}`)
+  const state = getState()
+
+  return processMarketResponse(dispatch, state, response)
+}
