+++ src/routes/MarketList/store/selectors/index.js
@@ -1,13 +1,13 @@
 import { List } from 'immutable'
 import { createSelector } from 'reselect'
 import { isMarketClosed, isMarketEndingSoon, isNewMarket } from 'store/utils/marketStatus'
-import { marketSelector } from 'store/selectors/market'
+import { marketsSelector } from 'store/selectors/market'
 
 import filterSelector from './filter'
 import sorterSelector from './sorter'
 
 export const marketListSelector = createSelector(
-  marketSelector,
+  marketsSelector,
   filterSelector,
   sorterSelector,
   (markets, filter, sorter) => List(markets.filter(filter).sort(sorter).values()),
