+++ src/routes/MarketList/store/test/market.reducer.js
@@ -3,7 +3,7 @@ import thunk from 'redux-thunk'
 import { List } from 'immutable'
 import marketReducer from 'store/reducers/market'
 import blockchainReducer from 'store/reducers/blockchain'
-import { processMarketResponse } from 'store/actions/market/fetchMarkets'
+import { processMarketsResponse } from 'store/actions/market/fetchMarkets'
 import { marketListSelector } from '../selectors'
 import { oneMarketData, twoMarketData, realData, MarketFactory } from './builder/index.builder'
 
@@ -30,7 +30,7 @@ const marketReducerTests = () => {
       const emptyResponse = { }
 
       // WHEN
-      processMarketResponse(store.dispatch, store.getState(), emptyResponse)
+      processMarketsResponse(store.dispatch, store.getState(), emptyResponse)
 
       // THEN
       const emptyList = List([])
@@ -43,7 +43,7 @@ const marketReducerTests = () => {
       const threeMarketsResponse = realData
 
       // WHEN
-      processMarketResponse(store.dispatch, store.getState(), threeMarketsResponse)
+      processMarketsResponse(store.dispatch, store.getState(), threeMarketsResponse)
 
       // THEN
       const markets = marketListSelector(store.getState())
@@ -60,11 +60,11 @@ const marketReducerTests = () => {
     it('should replace markets in store when fetch data', () => {
       // GIVEN
       const kittiesResponse = oneMarketData
-      processMarketResponse(store.dispatch, store.getState(), kittiesResponse)
+      processMarketsResponse(store.dispatch, store.getState(), kittiesResponse)
 
       // WHEN
       const etherAndGasMarketsResponse = twoMarketData
-      processMarketResponse(store.dispatch, store.getState(), etherAndGasMarketsResponse)
+      processMarketsResponse(store.dispatch, store.getState(), etherAndGasMarketsResponse)
 
       // THEN
       const markets = marketListSelector(store.getState())
