+++ src/routes/MarketList/store/test/market.selector.js
@@ -2,7 +2,7 @@ import { List } from 'immutable'
 import { marketListSelector } from '../selectors'
 
 const marketTests = () => {
-  describe('Market List Selector[marketSelector]', () => {
+  describe('Market List Selector[marketsSelector]', () => {
     it('should get empty immutable List when they are not loaded', () => {
       // GIVEN
       const emptyList = List([])
