+++ src/store/reducers/accountShares.js
